requirements.txt
psutil
pyyaml
requests
rich
